fastapi/requirements.txt
# FastAPI
# uvicorn>=0.15.0,<0.16.0
# fastapi>=0.68.0,<0.69.0
gunicorn==21.2.0
uvicorn==0.15.0
fastapi==0.109.2

# Middlewares
asgi-correlation-id==4.3.0
starlette-csrf==3.0.0
starlette-prometheus==0.9.0

# Extras
# pydantic>=1.8.0,<2.0.0
pydantic==2.6.1
pydantic_core==2.16.2
pytz==2024.1
psutil==5.9.8
itsdangerous==2.1.2
numpy==1.24.4
async_timeout==4.0.3
pyyaml==6.0.1

# External Services
elasticsearch==8.12.0
